Move type entity to dedicated folder
