+++ src/app/app-data/app-data.ts
@@ -1,6 +1,6 @@
 import { ItemData } from './items/item.data';
 import { InMemoryDbService } from 'angular-in-memory-web-api';
-import { ItemT } from './items/item.type';
+import { ItemT } from '../entities/item.type';
 
 export class AppData implements InMemoryDbService {
   public createDb(): { items: Array<ItemT> } {
