+++ src/app/app-data/items/item.service.ts
@@ -1,6 +1,6 @@
 import { combineLatest, map, Observable, Subject } from 'rxjs';
 import { inject, Injectable } from '@angular/core';
-import { ItemT, ItemIdT } from './item.type';
+import { ItemT, ItemIdT } from '../../entities/item.type';
 import { HttpClient } from '@angular/common/http';
 @Injectable({
   providedIn: 'root',
