--- src/app/app-data/items/item.type.ts
@@ -1,10 +1,0 @@
-export type ItemT = {
-  id: string;
-  title: string;
-  prices: { [tag: string]: number };
-  photos: string[];
-  description: string;
-  offerDiscount?: number;
-};
-
-export type ItemIdT = Pick<ItemT, 'id' | 'title'>;
